+++ docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGES
 
+## 1.11
+
+* Switched verification to using internal events
+* Added a recipe for creating a [private relay](../../../../wiki?name=Homeserver)
+
 ## 1.10.5
 
 * Add a `MultiIndex` to chain index results, reducing query misses.